refactor(db): consolidate shared repository helpers
