+++ packages/db/src/index.ts
@@ -2,4 +2,6 @@ export * from "./client";
 export * from "./repositories/admin.repo";
 export * from "./repositories/anime.repo";
 export * from "./repositories/library.repo";
+export * from "./repositories/shared/follow-counts";
+export * from "./repositories/shared/user-summary";
 export * from "./repositories/users.repo";
