+++ packages/db/src/repositories/admin.repo.ts
@@ -1,16 +1,11 @@
 import { db } from "../client";
-import { heroCuration, userFollow, userProfile } from "../schema/anilog";
+import { heroCuration, userProfile } from "../schema/anilog";
 import { user } from "../schema/auth";
-import { asc, count, eq, getTableColumns, ilike, inArray, or, sql } from "drizzle-orm";
+import { asc, count, eq, ilike, or } from "drizzle-orm";
+import { getFollowCountsMap } from "./shared/follow-counts";
+import { userSummarySelect, type UserSummaryRecord } from "./shared/user-summary";
 
-export type AdminUserRecord = {
-  id: string;
-  name: string;
-  username: string | null;
-  email: string;
-  isAdmin: boolean;
-  image: string | null;
-  profile: typeof userProfile.$inferSelect | null;
+export type AdminUserRecord = UserSummaryRecord & {
   followerCount: number;
   followingCount: number;
 };
@@ -27,61 +22,6 @@ type HeroCurationUpdateInput = Pick<
   | "sortOrder"
   | "isActive"
 >;
-
-function adminUserSelect() {
-  return {
-    id: user.id,
-    name: user.name,
-    username: user.username,
-    email: user.email,
-    isAdmin: user.isAdmin,
-    image: user.image,
-    profile: getTableColumns(userProfile),
-  };
-}
-
-async function getFollowCountsMap(userIds: string[]) {
-  if (userIds.length === 0) {
-    return new Map<string, { followerCount: number; followingCount: number }>();
-  }
-
-  const [followerRows, followingRows] = await Promise.all([
-    db
-      .select({
-        userId: userFollow.followingId,
-        count: sql<number>`count(*)::int`,
-      })
-      .from(userFollow)
-      .where(inArray(userFollow.followingId, userIds))
-      .groupBy(userFollow.followingId),
-    db
-      .select({
-        userId: userFollow.followerId,
-        count: sql<number>`count(*)::int`,
-      })
-      .from(userFollow)
-      .where(inArray(userFollow.followerId, userIds))
-      .groupBy(userFollow.followerId),
-  ]);
-
-  const counts = new Map<string, { followerCount: number; followingCount: number }>();
-
-  for (const userId of userIds) {
-    counts.set(userId, { followerCount: 0, followingCount: 0 });
-  }
-
-  for (const row of followerRows) {
-    const current = counts.get(row.userId) ?? { followerCount: 0, followingCount: 0 };
-    counts.set(row.userId, { ...current, followerCount: row.count });
-  }
-
-  for (const row of followingRows) {
-    const current = counts.get(row.userId) ?? { followerCount: 0, followingCount: 0 };
-    counts.set(row.userId, { ...current, followingCount: row.count });
-  }
-
-  return counts;
-}
 
 export class AdminRepository {
   static async getTotalUsers() {
@@ -101,7 +41,7 @@ export class AdminRepository {
       : undefined;
 
     const rows = await db
-      .select(adminUserSelect())
+      .select(userSummarySelect())
       .from(user)
       .leftJoin(userProfile, eq(user.id, userProfile.userId))
       .where(whereClause)
