+++ packages/domain/src/users/user.types.ts
@@ -1,9 +1,5 @@
 import type { PublicLibraryEntryDto, UserProfileDto, UserWithProfileDto } from "@anilog/contracts";
-import type {
-  PublicLibraryEntryRecord,
-  UserProfileRecord,
-  UserWithCountsRecord,
-} from "@anilog/db/repositories/users.repo";
+import type { PublicLibraryEntryRecord, UserProfileRecord, UserWithCountsRecord } from "@anilog/db";
 
 export type UserProfile = UserProfileRecord;
 export type UserProfileDtoShape = UserProfileDto;
